+++ test/firestore_ui_test.dart
@@ -1,15 +1,20 @@
 import 'dart:async';
 
 import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:firebase_core/firebase_core.dart'
-    show FirebaseApp, FirebaseOptions;
+import 'package:firebase_core/firebase_core.dart';
 import 'package:firestore_ui/firestore_list.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:cloud_firestore_platform_interface/src/method_channel/method_channel_firestore.dart';
+import 'package:firebase_core_platform_interface/firebase_core_platform_interface.dart';
 
 void main() {
+  TestWidgetsFlutterBinding.ensureInitialized();
+
   group('FirestoreList', () {
     int mockHandleId = 0;
+    final defaultBinaryMessenger =
+        ServicesBinding.instance.defaultBinaryMessenger;
     StreamController<QuerySnapshot> streamController;
     final List<MethodCall> log = <MethodCall>[];
     FirestoreList list;
@@ -21,8 +26,11 @@ void main() {
     };
 
     setUp(() async {
-      FirebaseApp.channel.setMockMethodCallHandler(
+      MethodChannelFirebaseCore.channel.setMockMethodCallHandler(
         (MethodCall methodCall) async {},
+      );
+      MethodChannelFirestore.channel.setMockMethodCallHandler(
+        (call) async {},
       );
 
       app = await FirebaseApp.configure(
@@ -43,7 +51,8 @@ void main() {
         debug: false,
       );
 
-      Firestore.channel.setMockMethodCallHandler((MethodCall methodCall) async {
+      MethodChannelFirestore.channel
+          .setMockMethodCallHandler((MethodCall methodCall) async {
         log.add(methodCall);
         switch (methodCall.method) {
           case 'Query#addSnapshotListener':
@@ -51,9 +60,9 @@ void main() {
             // Wait before sending a message back.
             // Otherwise the first request didn't have the time to finish.
             Future<void>.delayed(Duration.zero).then((_) {
-              BinaryMessages.handlePlatformMessage(
-                Firestore.channel.name,
-                Firestore.channel.codec.encodeMethodCall(
+              defaultBinaryMessenger.handlePlatformMessage(
+                MethodChannelFirestore.channel.name,
+                MethodChannelFirestore.channel.codec.encodeMethodCall(
                   MethodCall('QuerySnapshot', <String, dynamic>{
                     'app': app.name,
                     'handle': handle,
@@ -78,9 +87,9 @@ void main() {
             // Wait before sending a message back.
             // Otherwise the first request didn't have the time to finish.
             Future<void>.delayed(Duration.zero).then((_) {
-              BinaryMessages.handlePlatformMessage(
-                Firestore.channel.name,
-                Firestore.channel.codec.encodeMethodCall(
+              defaultBinaryMessenger.handlePlatformMessage(
+                MethodChannelFirestore.channel.name,
+                MethodChannelFirestore.channel.codec.encodeMethodCall(
                   MethodCall('DocumentSnapshot', <String, dynamic>{
                     'handle': handle,
                     'path': methodCall.arguments['path'],
@@ -153,5 +162,7 @@ void main() {
       expect(list.length, 1);
       expect(list[0].data, kMockDocumentSnapshotData);
     });
+
+    streamController?.close();
   });
 }